# docs/swagger/docs.json
{
  "swagger": "2.0",
  "info": {
    "contact": {
      "email": "hi@fanzru.dev",
      "name": "Social Media Service Team"
    },
    "description": "API for social media service",
    "title": "Social Media Service API",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/account/profile": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "401": {
            "description": "Unauthorized - invalid credentials",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Account"
        ],
        "description": "Get the profile of the authenticated user",
        "summary": "Get account profile"
      }
    },
...
    "/api/posts": {
      "get": {
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "description": "Cursor for pagination",
            "in": "query",
            "name": "cursor",
            "required": false,
            "type": "string"
          },
          {
            "default": 20,
            "description": "Number of posts to return (max 100)",
            "in": "query",
            "maximum": 100,
            "minimum": 1,
            "name": "limit",
            "required": false,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Posts retrieved successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "tags": [
          "Posts"
        ],
        "description": "Get all posts sorted by comment count with pagination",
        "summary": "Get all posts"
      },
      "post": {
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "maxLength": 1000,
            "minLength": 1,
            "name": "caption",
            "required": true,
            "type": "string"
          },
          {
            "description": "Image file (PNG, JPG, JPEG, BMP)",
            "format": "binary",
            "in": "formData",
            "name": "image",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Post created successfully",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "400": {
            "description": "Bad request - validation errors",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "401": {
            "description": "Unauthorized - invalid credentials",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "413": {
            "description": "Payload too large - image too big",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/StandardResponse"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Posts"
        ],
        "description": "Create a new social media post with image upload",
        "summary": "Create a new post"
      }
    },
...
  },
  "definitions": {
...
    "StandardResponse": {
      "properties": {
        "code": {
          "enum": [
            "SUCCESS",
            "FAILED",
            "BAD_REQUEST",
            "UNAUTHORIZED",
            "FORBIDDEN",
            "NOT_FOUND",
            "CONFLICT",
            "INTERNAL_SERVER_ERROR"
          ],
          "example": "SUCCESS",
          "type": "string"
        },
        "data": {
          "description": "Response data (varies by endpoint)",
          "type": "object"
        },
        "errors": {
          "example": [],
          "items": {
            "type": "string"
          },
          "type": "array"
        },
        "message": {
          "example": "Operation completed successfully",
          "type": "string"
        },
        "requestId": {
          "example": "req_123456789",
          "type": "string"
        },
        "serverTime": {
          "example": "2024-01-01T00:00:00Z",
          "format": "date-time",
          "type": "string"
        }
      },
      "type": "object"
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "description": "JWT token obtained from login endpoint",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
...
}