# internal/app/comment/port/genhttp/openapi_types.gen.go
// Package genhttp provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package genhttp

import (
	"time"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for StandardResponseCode.
const (
	BADREQUEST          StandardResponseCode = "BAD_REQUEST"
	CONFLICT            StandardResponseCode = "CONFLICT"
	FAILED              StandardResponseCode = "FAILED"
	FORBIDDEN           StandardResponseCode = "FORBIDDEN"
	INTERNALSERVERERROR StandardResponseCode = "INTERNAL_SERVER_ERROR"
	NOTFOUND            StandardResponseCode = "NOT_FOUND"
	SUCCESS             StandardResponseCode = "SUCCESS"
	UNAUTHORIZED        StandardResponseCode = "UNAUTHORIZED"
)

// CreateCommentRequest defines model for CreateCommentRequest.
type CreateCommentRequest struct {
	Content string `json:"content"`
}

// StandardResponse defines model for StandardResponse.
type StandardResponse struct {
	Code *StandardResponseCode `json:"code,omitempty"`

	// Data Response data (varies by endpoint)
	Data       *map[string]interface{} `json:"data,omitempty"`
	Errors     *[]string               `json:"errors,omitempty"`
	Message    *string                 `json:"message,omitempty"`
	RequestId  *string                 `json:"requestId,omitempty"`
	ServerTime *time.Time              `json:"serverTime,omitempty"`
}

// StandardResponseCode defines model for StandardResponse.Code.
type StandardResponseCode string

// UpdateCommentRequest defines model for UpdateCommentRequest.
type UpdateCommentRequest struct {
	Content string `json:"content"`
}

// GetApiCommentsUserUserIdParams defines parameters for GetApiCommentsUserUserId.
type GetApiCommentsUserUserIdParams struct {
	// Cursor Cursor for pagination
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// Limit Number of comments to return (max 100)
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// GetApiPostsPostIdCommentsParams defines parameters for GetApiPostsPostIdComments.
type GetApiPostsPostIdCommentsParams struct {
	// Cursor Cursor for pagination
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// Limit Number of comments to return (max 100)
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// PutApiCommentsIdJSONRequestBody defines body for PutApiCommentsId for application/json ContentType.
type PutApiCommentsIdJSONRequestBody = UpdateCommentRequest

// PostApiPostsPostIdCommentsJSONRequestBody defines body for PostApiPostsPostIdComments for application/json ContentType.
type PostApiPostsPostIdCommentsJSONRequestBody = CreateCommentRequest