# internal/app/post/port/genhttp/openapi_types.gen.go
// Package genhttp provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package genhttp

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// Defines values for StandardResponseCode.
const (
	BADREQUEST          StandardResponseCode = "BAD_REQUEST"
	CONFLICT            StandardResponseCode = "CONFLICT"
	FAILED              StandardResponseCode = "FAILED"
	FORBIDDEN           StandardResponseCode = "FORBIDDEN"
	INTERNALSERVERERROR StandardResponseCode = "INTERNAL_SERVER_ERROR"
	NOTFOUND            StandardResponseCode = "NOT_FOUND"
	SUCCESS             StandardResponseCode = "SUCCESS"
	UNAUTHORIZED        StandardResponseCode = "UNAUTHORIZED"
)

// StandardResponse defines model for StandardResponse.
type StandardResponse struct {
	Code *StandardResponseCode `json:"code,omitempty"`

	// Data Response data (varies by endpoint)
	Data       *map[string]interface{} `json:"data,omitempty"`
	Errors     *[]string               `json:"errors,omitempty"`
	Message    *string                 `json:"message,omitempty"`
	RequestId  *string                 `json:"requestId,omitempty"`
	ServerTime *time.Time              `json:"serverTime,omitempty"`
}

// StandardResponseCode defines model for StandardResponse.Code.
type StandardResponseCode string

// UpdatePostRequest defines model for UpdatePostRequest.
type UpdatePostRequest struct {
	Caption string `json:"caption"`
}

// GetApiPostsParams defines parameters for GetApiPosts.
type GetApiPostsParams struct {
	// Cursor Cursor for pagination
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// Limit Number of posts to return (max 100)
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// PostApiPostsMultipartBody defines parameters for PostApiPosts.
type PostApiPostsMultipartBody struct {
	Caption string `json:"caption"`

	// Image Image file (PNG, JPG, JPEG, BMP)
	Image openapi_types.File `json:"image"`
}

// GetApiPostsUserUserIdParams defines parameters for GetApiPostsUserUserId.
type GetApiPostsUserUserIdParams struct {
	// Cursor Cursor for pagination
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// Limit Number of posts to return (max 100)
	Limit *int `form:"limit,omitempty" json:"limit,omitempty"`
}

// PostApiPostsMultipartRequestBody defines body for PostApiPosts for multipart/form-data ContentType.
type PostApiPostsMultipartRequestBody PostApiPostsMultipartBody

// PutApiPostsIdJSONRequestBody defines body for PutApiPostsId for application/json ContentType.
type PutApiPostsIdJSONRequestBody = UpdatePostRequest